requirements.txt
Flask
connexion
Flask-Injector
fastavro
nameko
cassandra-driver
pika